moved tests to GoToFileFacts
